added NOOP delegate adapter and ability to build service with keystore
